# swagger.yaml
swagger: "2.0"
info:
  title: AvailabilityBackend API
  version: "0.1.0"
host: localhost:3000
basePath: /api
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: "Use: Bearer <token>"
paths:
  /auth/register:
    post:
      tags:
        - Auth
      summary: Register user
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/RegisterRequest"
      responses:
        201:
          description: Registered
          schema:
            $ref: "#/definitions/AuthResponse"
        400:
          description: Bad request
          schema:
            $ref: "#/definitions/ErrorResponse"
  /auth/login:
    post:
      tags:
        - Auth
      summary: Login user
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/LoginRequest"
      responses:
        200:
          description: Logged in
          schema:
            $ref: "#/definitions/AuthResponse"
        401:
          description: Invalid credentials
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
  RegisterRequest:
    type: object
    required:
      - email
      - password
    properties:
      email:
        type: string
        format: email
        example: student@example.com
      password:
        type: string
        example: password123
  LoginRequest:
    type: object
    required:
      - email
      - password
    properties:
      email:
        type: string
        format: email
        example: student@example.com
      password:
        type: string
        example: password123
  AuthResponse:
    type: object
    properties:
      token:
        type: string
        example: eyJhbGciOiJIUzI1...
      user:
        $ref: "#/definitions/User"
  User:
    type: object
    properties:
      id:
        type: string
        example: usr_1
      email:
        type: string
        example: student@example.com
...
  ErrorResponse:
    type: object
    properties:
      message:
        type: string
        example: Something went wrong